Support Presidio DI components as roots
If a Needle `Component` has a parent that is a `Presidio.Component`, an additional dependency provider needs to be generated, where the `Presidio.Component` acts as the root of the branch. This is because the `Presidio.Component` can be instantiated by any object even if that object is not on the DI graph. In that case, the `Presidio.Component` is the root of an entirely new DI tree.